static/margin.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-margin-block with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-margin-block">
<meta name="assert" content="scroll-margin-block supports only the grammar '<length>{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("scroll-margin-block-start", "auto");
test_invalid_value("scroll-margin-block-start", "20%");
test_invalid_value("scroll-margin-block-start", "-30%");
test_invalid_value("scroll-margin-block-start", "1px 2px");


test_invalid_value("scroll-margin-block-end", "auto");
test_invalid_value("scroll-margin-block-end", "20%");
test_invalid_value("scroll-margin-block-end", "-30%");
test_invalid_value("scroll-margin-block-end", "1px 2px");


test_invalid_value("scroll-margin-block", "auto");
test_invalid_value("scroll-margin-block", "20%");
test_invalid_value("scroll-margin-block", "-30%");
test_invalid_value("scroll-margin-block", "1px 2px 3px");








test_valid_value("scroll-margin-block-start", "-10px");
test_valid_value("scroll-margin-block-start", "calc(2em + 3ex)");


test_valid_value("scroll-margin-block-end", "-10px");
test_valid_value("scroll-margin-block-end", "calc(2em + 3ex)");


test_valid_value("scroll-margin-block", "-10px");
test_valid_value("scroll-margin-block", "calc(2em + 3ex)");

test_valid_value("scroll-margin-block", "1px 2px");











test_invalid_value("scroll-margin-inline-start", "auto");
test_invalid_value("scroll-margin-inline-start", "20%");
test_invalid_value("scroll-margin-inline-start", "-30%");
test_invalid_value("scroll-margin-inline-start", "1px 2px");


test_invalid_value("scroll-margin-inline-end", "auto");
test_invalid_value("scroll-margin-inline-end", "20%");
test_invalid_value("scroll-margin-inline-end", "-30%");
test_invalid_value("scroll-margin-inline-end", "1px 2px");


test_invalid_value("scroll-margin-inline", "auto");
test_invalid_value("scroll-margin-inline", "20%");
test_invalid_value("scroll-margin-inline", "-30%");
test_invalid_value("scroll-margin-inline", "1px 2px 3px");












test_valid_value("scroll-margin-inline-start", "-10px");
test_valid_value("scroll-margin-inline-start", "calc(2em + 3ex)");


test_valid_value("scroll-margin-inline-end", "-10px");
test_valid_value("scroll-margin-inline-end", "calc(2em + 3ex)");


test_valid_value("scroll-margin-inline", "-10px");
test_valid_value("scroll-margin-inline", "calc(2em + 3ex)");
test_valid_value("scroll-margin-inline", "1px 2px");














test_invalid_value("scroll-margin-top", "auto");
test_invalid_value("scroll-margin-top", "20%");
test_invalid_value("scroll-margin-top", "-30%");
test_invalid_value("scroll-margin-top", "1px 2px");


test_invalid_value("scroll-margin-right", "auto");
test_invalid_value("scroll-margin-right", "20%");
test_invalid_value("scroll-margin-right", "-30%");
test_invalid_value("scroll-margin-right", "1px 2px");


test_invalid_value("scroll-margin-bottom", "auto");
test_invalid_value("scroll-margin-bottom", "20%");
test_invalid_value("scroll-margin-bottom", "-30%");
test_invalid_value("scroll-margin-bottom", "1px 2px");


test_invalid_value("scroll-margin-left", "auto");
test_invalid_value("scroll-margin-left", "20%");
test_invalid_value("scroll-margin-left", "-30%");
test_invalid_value("scroll-margin-left", "1px 2px");


test_invalid_value("scroll-margin", "auto");
test_invalid_value("scroll-margin", "20%");
test_invalid_value("scroll-margin", "-30%");
test_invalid_value("scroll-margin", "1px 2px 3px 4px 5px");












test_valid_value("scroll-margin-top", "-10px");
test_valid_value("scroll-margin-top", "calc(2em + 3ex)");


test_valid_value("scroll-margin-right", "-10px");
test_valid_value("scroll-margin-right", "calc(2em + 3ex)");


test_valid_value("scroll-margin-bottom", "-10px");
test_valid_value("scroll-margin-bottom", "calc(2em + 3ex)");


test_valid_value("scroll-margin-left", "-10px");
test_valid_value("scroll-margin-left", "calc(2em + 3ex)");


test_valid_value("scroll-margin", "-10px");
test_valid_value("scroll-margin", "calc(2em + 3ex)");

test_valid_value("scroll-margin", "1px 2px");
test_valid_value("scroll-margin", "1px 2px 3px");
test_valid_value("scroll-margin", "1px 2px 3px 4px");
</script>
</body>
</html>
